+++ packages/bpk-tokens/tokens/base.react.native.android.js
@@ -1,0 +1,275 @@
+/*
+ * 
+ * Backpack - Skyscanner's Design System
+ * 
+ * Copyright 2017 Skyscanner Ltd
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * 
+ */
+export const animationDurationXs = "50ms";
+export const animationDurationSm = "200ms";
+export const animationDurationBase = "400ms";
+export const borderSizeSm = 1;
+export const borderSizeLg = 2;
+export const borderSizeXl = 3;
+export const borderRadiusSm = 2;
+export const borderRadiusPill = 40;
+export const buttonBorderWidth = 0;
+export const buttonHeight = 32 + 4;
+export const buttonPaddingHorizontal = 4 * 3;
+export const buttonLineHeightLarge = 4 * 12;
+export const buttonPaddingVertical = 8;
+export const colorWhite = "rgb(255, 255, 255)";
+export const colorBlue50 = "rgb(225, 244, 248)";
+export const colorBlue100 = "rgb(203, 238, 245)";
+export const colorBlue200 = "rgb(176, 228, 238)";
+export const colorBlue300 = "rgb(127, 215, 232)";
+export const colorBlue400 = "rgb(64, 196, 223)";
+export const colorBlue500 = "rgb(0, 178, 214)";
+export const colorBlue600 = "rgb(0, 157, 189)";
+export const colorBlue700 = "rgb(0, 140, 168)";
+export const colorBlue800 = "rgb(0, 117, 140)";
+export const colorBlue900 = "rgb(0, 85, 103)";
+export const colorGray50 = "rgb(243, 242, 245)";
+export const colorGray100 = "rgb(230, 228, 235)";
+export const colorGray200 = "rgb(204, 201, 212)";
+export const colorGray300 = "rgb(178, 174, 189)";
+export const colorGray400 = "rgb(154, 149, 167)";
+export const colorGray500 = "rgb(129, 123, 143)";
+export const colorGray600 = "rgb(105, 97, 121)";
+export const colorGray700 = "rgb(82, 76, 97)";
+export const colorGray800 = "rgb(59, 52, 75)";
+export const colorGray900 = "rgb(37, 32, 51)";
+export const colorGreen50 = "rgb(223, 247, 236)";
+export const colorGreen100 = "rgb(203, 245, 226)";
+export const colorGreen200 = "rgb(175, 237, 209)";
+export const colorGreen300 = "rgb(128, 232, 185)";
+export const colorGreen400 = "rgb(64, 222, 151)";
+export const colorGreen500 = "rgb(0, 215, 117)";
+export const colorGreen600 = "rgb(0, 189, 104)";
+export const colorGreen700 = "rgb(0, 168, 93)";
+export const colorGreen800 = "rgb(0, 140, 77)";
+export const colorGreen900 = "rgb(0, 102, 56)";
+export const colorRed50 = "rgb(252, 242, 242)";
+export const colorRed100 = "rgb(255, 214, 213)";
+export const colorRed200 = "rgb(255, 187, 186)";
+export const colorRed300 = "rgb(255, 150, 148)";
+export const colorRed400 = "rgb(254, 116, 113)";
+export const colorRed500 = "rgb(255, 84, 82)";
+export const colorRed600 = "rgb(235, 66, 63)";
+export const colorRed700 = "rgb(222, 50, 47)";
+export const colorRed800 = "rgb(204, 31, 29)";
+export const colorRed900 = "rgb(168, 3, 0)";
+export const colorYellow50 = "rgb(255, 249, 230)";
+export const colorYellow100 = "rgb(255, 243, 207)";
+export const colorYellow200 = "rgb(255, 236, 184)";
+export const colorYellow300 = "rgb(255, 225, 140)";
+export const colorYellow400 = "rgb(255, 207, 74)";
+export const colorYellow500 = "rgb(255, 187, 0)";
+export const colorYellow600 = "rgb(240, 176, 0)";
+export const colorYellow700 = "rgb(225, 165, 0)";
+export const colorYellow800 = "rgb(194, 142, 0)";
+export const colorYellow900 = "rgb(156, 114, 0)";
+export const colorPink50 = "rgb(253, 228, 237)";
+export const colorPink100 = "rgb(255, 191, 215)";
+export const colorPink200 = "rgb(255, 148, 187)";
+export const colorPink300 = "rgb(255, 115, 166)";
+export const colorPink400 = "rgb(255, 97, 155)";
+export const colorPink500 = "rgb(250, 72, 138)";
+export const colorPink600 = "rgb(217, 43, 107)";
+export const colorPink700 = "rgb(197, 15, 82)";
+export const colorPink800 = "rgb(176, 12, 72)";
+export const colorPink900 = "rgb(148, 5, 58)";
+export const elevationXs = 2;
+export const elevationSm = 4;
+export const elevationBase = 6;
+export const elevationLg = 16;
+export const elevationXl = 24;
+export const spacingSm = 4;
+export const spacingMd = 8;
+export const spacingBase = 16;
+export const spacingLg = 24;
+export const spacingXl = 32;
+export const spacingXxl = 40;
+export const fontFamily = "sans-serif";
+export const fontFamilyEmphasize = "sans-serif-medium";
+export const fontSizeXs = 12;
+export const fontSizeSm = 14;
+export const fontSizeBase = 16;
+export const fontSizeLg = 20;
+export const fontSizeXl = 24;
+export const fontSizeXxl = 34;
+export const lineHeightXs = 18;
+export const lineHeightSm = 20;
+export const lineHeightBase = 24;
+export const lineHeightLg = 28;
+export const lineHeightXl = 32;
+export const lineHeightXxl = 40;
+export const textXsFontSize = 12;
+export const textXsFontWeight = "400";
+export const textXsLineHeight = 18;
+export const textSmFontSize = 14;
+export const textSmFontWeight = "400";
+export const textSmLineHeight = 20;
+export const textBaseFontSize = 16;
+export const textBaseFontWeight = "400";
+export const textBaseLineHeight = 24;
+export const textLgFontSize = 20;
+export const textLgFontWeight = "400";
+export const textLgLineHeight = 28;
+export const textXlFontSize = 24;
+export const textXlFontWeight = "400";
+export const textXlLineHeight = 32;
+export const textXxlFontSize = 34;
+export const textXxlFontWeight = "400";
+export const textXxlLineHeight = 40;
+export const textEmphasizedFontWeight = "500";
+export const animations = {
+animationDurationXs,
+animationDurationSm,
+animationDurationBase,
+};
+export const borders = {
+borderSizeSm,
+borderSizeLg,
+borderSizeXl,
+};
+export const radii = {
+borderRadiusSm,
+borderRadiusPill,
+};
+export const spacings = {
+buttonBorderWidth,
+buttonHeight,
+buttonPaddingHorizontal,
+buttonLineHeightLarge,
+buttonPaddingVertical,
+spacingSm,
+spacingMd,
+spacingBase,
+spacingLg,
+spacingXl,
+spacingXxl,
+};
+export const colors = {
+colorWhite,
+colorBlue50,
+colorBlue100,
+colorBlue200,
+colorBlue300,
+colorBlue400,
+colorBlue500,
+colorBlue600,
+colorBlue700,
+colorBlue800,
+colorBlue900,
+colorGray50,
+colorGray100,
+colorGray200,
+colorGray300,
+colorGray400,
+colorGray500,
+colorGray600,
+colorGray700,
+colorGray800,
+colorGray900,
+colorGreen50,
+colorGreen100,
+colorGreen200,
+colorGreen300,
+colorGreen400,
+colorGreen500,
+colorGreen600,
+colorGreen700,
+colorGreen800,
+colorGreen900,
+colorRed50,
+colorRed100,
+colorRed200,
+colorRed300,
+colorRed400,
+colorRed500,
+colorRed600,
+colorRed700,
+colorRed800,
+colorRed900,
+colorYellow50,
+colorYellow100,
+colorYellow200,
+colorYellow300,
+colorYellow400,
+colorYellow500,
+colorYellow600,
+colorYellow700,
+colorYellow800,
+colorYellow900,
+colorPink50,
+colorPink100,
+colorPink200,
+colorPink300,
+colorPink400,
+colorPink500,
+colorPink600,
+colorPink700,
+colorPink800,
+colorPink900,
+};
+export const elevation = {
+elevationXs,
+elevationSm,
+elevationBase,
+elevationLg,
+elevationXl,
+};
+export const typesettings = {
+fontFamily,
+fontFamilyEmphasize,
+fontSizeXs,
+fontSizeSm,
+fontSizeBase,
+fontSizeLg,
+fontSizeXl,
+fontSizeXxl,
+lineHeightXs,
+lineHeightSm,
+lineHeightBase,
+lineHeightLg,
+lineHeightXl,
+lineHeightXxl,
+};
+export const fontSizes = {
+textXsFontSize,
+textSmFontSize,
+textBaseFontSize,
+textLgFontSize,
+textXlFontSize,
+textXxlFontSize,
+};
+export const fontWeights = {
+textXsFontWeight,
+textSmFontWeight,
+textBaseFontWeight,
+textLgFontWeight,
+textXlFontWeight,
+textXxlFontWeight,
+textEmphasizedFontWeight,
+};
+export const lineHeights = {
+textXsLineHeight,
+textSmLineHeight,
+textBaseLineHeight,
+textLgLineHeight,
+textXlLineHeight,
+textXxlLineHeight,
+};